Add files via upload
I wrote this script a few years back to look up driving distance between two addresses (via Google/GIS) for someone who was working on a grad project, and it was for a good cause in general. It worked well at the time, but might need some mods to work today. Some of the code snippets might be useful.